rename theme provider files and class
